flask-app/requirements.txt
Flask==2.1.1
Werkzeug==2.0.1
psycopg2-binary==2.9.5
redis==4.5.4
prometheus-flask-exporter==0.22.4